Add skip dependency check param to Azure Pipelines
Introduced a new `SkipDependencyCheck` parameter in the Azure Pipelines configuration to allow users the option to skip the dependency check step. This parameter defaults to `false`, ensuring the dependency check is performed unless explicitly bypassed. Additionally, modified the `PERFORM_DEPENDENCYCHECK_DOTNETBUILD_DOTNETTEST_AND_PUBLISH` stage to include a condition that checks the value of `SkipDependencyCheck`. If set to `true`, the `Dependencychecker` job is skipped, facilitating more flexible pipeline executions.