Add step to restore GitHub submodules in CI
This commit updates the `continuous-testing.yml` file by adding a step to restore GitHub submodules using the command `git submodule update --init --recursive`. This step is positioned after the installation of Playwright dependencies and specifies the working directory as `$(Build.SourcesDirectory)`. A comment has also been included to express uncertainty about the correctness of this change.